--- lib/Exception/BlindIndexNameCollisionException.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const CipherSweetException = require('./CipherSweetException');
-
-/**
- * @class   BlindIndexNameCollisionException
- * @package CipherSweet.Exception
- * @author  Paragon Initiative Enterprises
- */
-module.exports = class BlindIndexNameCollisionException extends CipherSweetException {};
