--- lib/Exception/BlindIndexNotFoundException.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const CipherSweetException = require('./CipherSweetException');
-
-/**
- * @class   BlindIndexNotFoundException
- * @package CipherSweet.Exception
- * @author  Paragon Initiative Enterprises
- */
-module.exports = class BlindIndexNotFoundException extends CipherSweetException {};
