--- lib/Exception/CipherSweetException.js
@@ -1,8 +1,0 @@
-"use strict";
-
-/**
- * @class   CipherSweetException
- * @package CipherSweet.Exception
- * @author  Paragon Initiative Enterprises
- */
-module.exports = class CipherSweetException extends Error {};
