--- lib/Exception/CryptoOperationException.js
@@ -1,10 +1,0 @@
-"use strict";
-
-const CipherSweetException = require('./CipherSweetException');
-
-/**
- * @class   CryptoOperationException
- * @package CipherSweet.Exception
- * @author  Paragon Initiative Enterprises
- */
-module.exports = class CryptoOperationException extends CipherSweetException {};
