+++ sync-lib/Backend/Key/SymmetricKey.js
@@ -1,0 +1,37 @@
+"use strict";
+
+/**
+ * @class   CipherSweet
+ * @package CipherSweet.Backend.Key
+ * @author  Paragon Initiative Enterprises
+ */
+module.exports = class SymmetricKey
+{
+    /**
+     * @param {string|Buffer} rawKeyMaterial
+     */
+    constructor(rawKeyMaterial)
+    {
+        if (!Buffer.isBuffer(rawKeyMaterial)) {
+            rawKeyMaterial = Buffer.from(rawKeyMaterial, 'binary');
+        }
+        this.rawKeyMaterial = rawKeyMaterial;
+    }
+
+    /**
+     * @param {object} key
+     * @return {boolean}
+     */
+    static isSymmetricKey(key)
+    {
+        return Buffer.isBuffer(key.rawKeyMaterial);
+    }
+
+    /**
+     * @returns {Buffer}
+     */
+    getRawKey()
+    {
+        return this.rawKeyMaterial;
+    }
+};
