+++ sync-lib/CompoundIndex.js
@@ -1,0 +1,141 @@
+"use strict";
+
+const Compound = require('./Transformation/Compound');
+const Transformation = require('./Contract/Transformation');
+const Util = require('./Util');
+
+/**
+ * Class CompoundIndex
+ *
+ * @package CipherSweet
+ */
+module.exports = class CompoundIndex
+{
+    /**
+     *
+     * @param {string} name
+     * @param {Array<int, string>} columns
+     * @param {Number} filterBits
+     * @param {boolean} fastHash
+     * @param {Object<string, int>} hashConfig
+     */
+    constructor(name, columns = [], filterBits = 256, fastHash = false, hashConfig = {})
+    {
+        this.name = name;
+        this.columns = columns;
+        this.filterBits = filterBits;
+        this.fastHash = fastHash;
+        this.hashConfig = hashConfig;
+        this.columnTransforms = {};
+        this.rowTransforms = [];
+    }
+
+    /**
+     * @param {string} column
+     * @param {module.Transformation} tf
+     * @return {module.CompoundIndex}
+     */
+    addTransform(column, tf)
+    {
+        if (!this.columnTransforms[column]) {
+            this.columnTransforms[column] = [];
+        }
+        this.columnTransforms[column].push(tf);
+        return this;
+    }
+
+    /**
+     * @param {module.RowTransformation} tf
+     */
+    addRowTransform(tf)
+    {
+        this.rowTransforms.push(tf);
+    }
+
+    /**
+     * @returns {Array<int, string>}
+     */
+    getColumns()
+    {
+        return this.columns;
+    }
+
+    /**
+     * @returns {string}
+     */
+    getName()
+    {
+        return this.name;
+    }
+
+    /**
+     * @returns {boolean}
+     */
+    getFastHash()
+    {
+        return this.fastHash;
+    }
+
+    /**
+     * @returns {Number}
+     */
+    getFilterBitLength()
+    {
+        return this.filterBits;
+    }
+
+    /**
+     * @returns {Object<string, int>}
+     */
+    getHashConfig()
+    {
+        return this.hashConfig;
+    }
+
+    /**
+     * @returns {module.RowTransformation[]}
+     */
+    getRowTransforms()
+    {
+        return this.rowTransforms;
+    }
+
+    /**
+     * @param {Array} row
+     * @return {string}
+     */
+    getPacked(row)
+    {
+        let col;
+        let piece;
+        let pieces = {};
+        let tf;
+        for (let i = 0; i < this.columns.length; i++) {
+            col = this.columns[i];
+            if (typeof row[col] === 'undefined') {
+                continue;
+            }
+            piece = row[col];
+            if (this.columnTransforms[col]) {
+                /** @var {string} t */
+                for (let t = 0; t < this.columnTransforms[col].length; t++) {
+                    tf = this.columnTransforms[col][t];
+                    piece = tf.invoke(piece);
+                }
+            }
+            pieces[col] = piece;
+        }
+
+        if (this.rowTransforms.length > 0) {
+            for (let t = 0; t < this.rowTransforms[col].length; t++) {
+                tf = this.rowTransforms[t];
+                pieces = tf(pieces);
+            }
+        }
+
+        if (typeof pieces === 'string') {
+            return pieces;
+        }
+        return (new Compound()).invoke(pieces);
+    }
+};
