+++ sync-lib/Exception/ArrayKeyException.js
@@ -1,0 +1,10 @@
+"use strict";
+
+const CipherSweetException = require('./CipherSweetException');
+
+/**
+ * @class   ArrayKeyException
+ * @package CipherSweet.Exception
+ * @author  Paragon Initiative Enterprises
+ */
+module.exports = class ArrayKeyException extends CipherSweetException {};
