+++ sync-lib/Exception/CipherSweetException.js
@@ -1,0 +1,8 @@
+"use strict";
+
+/**
+ * @class   CipherSweetException
+ * @package CipherSweet.Exception
+ * @author  Paragon Initiative Enterprises
+ */
+module.exports = class CipherSweetException extends Error {};
